requirements.txt
appnope==0.1.0
cling==2.3.1
coverage==4.0.3
coveralls==1.1
decorator==4.0.6
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.1
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
docopt==0.6.2
gnureadline==6.3.3
gunicorn==19.4.5
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
PyJWT==1.4.0
PyYAML==3.11
requests==2.9.1
simplegeneric==0.8.1
static3==0.6.1
traitlets==4.0.0
travis==0.0.2
wheel==0.24.0
whitenoise==2.0.6
